+++ django_component_kit/__init__.py
@@ -1,6 +1,6 @@
 from .tags import do_merge_attrs, do_slot, do_render_slot, do_partial
 from .decorators import component_inline_tag, component_block_tag
-from .utils import render_partial_from_template
+from .partials import render_partial_from_template
 
 __all__ = [
     "do_merge_attrs",